Reworked searching for posts so it can use arguments.
Updated the searching capabilites from what it was to now using the '/posts/search?key1=val1&key2=val2..." format. This allows for cleaner and variable urls. It also allows for easier searching and filtering by user or hall or meal(and anything else needed in the future) at the same time. It can also be used as a filter by omitting a search key and just utilizing the filtering options.