Added a "match" option to Search
"Match" option lets the user define how the key is found in the posts, by 3 options, contains for the key appearing in any form in the post, starts for the key appearing at the start of words in the post, and the default, is, which will only match the key if it is the whole word in the post. This search option can be used using parameter "match=option" where option is on the of the stated options.