Complete overhaul of the Frontend Search Feature
The frontend search feature has been changed to appear very similar but acts in a more robust and productive way. The old model was search while inputting specific filters in the search field. Now switched to using drop downs to select filters and then only searching for keywords. Still operates the same back end wise(subject to change). Notably however now query parameters are being used instead of path parameters in compliance with the backend model for searchs. Right now it is not very well styled but it is at the very least functional. There might be some errors, please feel free to fix them or let me know, also feel free to do however you wish with the styling and such. I have kept the old model in for now in case this new way finds a way to blow itself up.