Updated Search bar on the Searched page
Search bar on the Searched Page now takes the form of a SearchAndFilterBar(admittedly not my best name) which allows for setting filter options along with sending another search after just searching. Right now the filter options will preserve through a search but the actual search keyword will not be if coming from the homepage. This might be desired but seems sort of janky.